Restructure/rename NavHeader/Header components, update pages
