--- frontend/components/Header/index.tsx
@@ -1,4 +1,0 @@
-import Header from "./Header";
-import NavHeader from "./NavHeader";
-
-export { Header, NavHeader };
