+++ frontend/components/LoginHeader/index.tsx
@@ -1,0 +1,3 @@
+import LoginHeader from "./LoginHeader";
+
+export { LoginHeader };
