+++ frontend/components/NavHeader/index.tsx
@@ -1,5 +1,6 @@
+import NavHeader from "./NavHeader";
 import NavList from "./NavList";
 import { NavMenuButton } from "./NavMenuButton";
 import NavMenuListItem from "./NavMenuListItem";
 
-export { NavList, NavMenuButton, NavMenuListItem };
+export { NavList, NavHeader, NavMenuButton, NavMenuListItem };
