+++ frontend/pages/admin/create-workspace.tsx
@@ -7,7 +7,7 @@ import { useForm } from "react-hook-form";
 import styled from "styled-components";
 
 import { Head } from "../../components/Head";
-import { Header } from "../../components/Header";
+import { LoginHeader } from "../../components/LoginHeader";
 import { MainHeading } from "../../components/MainHeading";
 import { PageLayout } from "../../components/PageLayout";
 import { Textarea } from "../../components/Textarea";
@@ -96,7 +96,7 @@ const CreateWorkspace = () => {
     <>
       <Head title="Admin - Create Workspace" />
       <PageLayout>
-        <Header />
+        <LoginHeader />
         <PageContent>
           <MainHeading>Create a workspace</MainHeading>
           <H2>Workspace details</H2>
