+++ frontend/pages/workspaces/pharmacy.tsx
@@ -3,7 +3,7 @@ import React from "react";
 import { GetServerSideProps } from "next";
 import Link from "next/link";
 
-import { Header } from "../../components/Header";
+import { LoginHeader } from "../../components/LoginHeader";
 import { MainHeading } from "../../components/MainHeading";
 import { requireAuthentication } from "../../lib/auth";
 
@@ -18,7 +18,7 @@ export const getServerSideProps: GetServerSideProps = requireAuthentication(
 const PrivatePage = () => {
   return (
     <>
-      <Header />
+      <LoginHeader />
       <MainHeading>Pharmacy Page</MainHeading>
       <p>This is a different private page</p>
       <Link href="/workspaces/private">
